Move navigateToSearchedItem to separate server file to avoid conflicts with client components
