+++ src/components/navigation/utilities-nav-menu.tsx
@@ -1,10 +1,8 @@
 import { getCartCount } from '@/lib/actions/cart';
-import { routes } from '@/lib/constants/routes';
-import { getSlugFromTitle } from '@/lib/data/helpers';
+import { navigateToSearchedItem } from '@/lib/actions/search';
 import { getProductsLight } from '@/lib/data/product-data-service';
 import { Result } from '@/lib/types/types';
 import { ShoppingCart } from 'lucide-react';
-import { redirect } from 'next/navigation';
 import { Suspense } from 'react';
 import { SearchBar } from '../search/search-bar';
 import {
@@ -24,16 +22,6 @@ export async function UtilitiesNavMenu({ className }: { className?: string }) {
       const errorMsg = typeof error === 'string' ? error : '';
       return { success: false, error: errorMsg };
     }
-  };
-
-  const navigateToSearchedItem = async (formData: FormData) => {
-    'use server';
-    const searchTerm = formData.get('search') as string;
-    if (!searchTerm) return;
-
-    const slug = getSlugFromTitle(searchTerm);
-
-    redirect(`${routes.products.href}/${slug}`);
   };
 
   return (
